tiling编译问题处理

+++ mc2/grouped_mat_mul_allto_allv/op_host/op_tiling/arch35/quant_grouped_mat_mul_allto_allv_tiling_adapter.cpp
@@ -187,7 +187,7 @@ using namespace optiling;
 //     return ge::GRAPH_SUCCESS;
 // }
 
-ge::graphStatus QuantGroupedMatmulAllToAllvAdapter::SetExpertInputParameters(const int64_t* sendCounts,
+ge::graphStatus QuantGroupedMatmulAllToAllvAdapter::SetExpertInputParameters(const int16_t* sendCounts,
     uint64_t worldSize, uint64_t index, uint32_t epNums)
 {
     // uint32_t worldSize = tilingProcesser_.localTilingData_.taskTilingInfo.epWorldSize;
